executor: migrate test-infra to testify for part3 of executor_test (#32884)
ref pingcap/tidb#28577
